src/workers/requirements.txt
pydantic==2.1.1
pydantic-settings==2.0.2
python-dotenv
pytest
filelock==3.13.1
fsspec==2024.3.1
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
Pillow==8.1.0
sympy==1.12
timm==0.3.2
torch==1.7.0
torchaudio==0.7.0
torchvision==0.8.0
pyyaml
python-multipart
pika==1.3.2
aio_pika==9.4.1
redis==5.0.3
onnx
onnxscript
onnxruntime
protobuf==3.20.*